improve tui code structure and details
