+++ src/main.rs
@@ -1,50 +1,36 @@
 mod samples;
+mod tui;
 
-use std::fs::{File, metadata};
-use std::io::{BufRead, BufReader};
-use std::time::{Instant, Duration}; // , SystemTime, UNIX_EPOCH
+use tui::*;
+
 use std::error::Error;
-use std::{io, thread}; // use std::time::{Instant};
+use std::fs::{metadata, File};
+use std::io::{BufRead, BufReader};
+use std::sync::atomic::{AtomicBool, AtomicPtr, AtomicU32, AtomicUsize, Ordering};
 use std::sync::Arc;
-use std::sync::atomic::{AtomicUsize, AtomicU32, AtomicBool, AtomicPtr, Ordering};
+use std::time::{Duration, Instant}; // , SystemTime, UNIX_EPOCH
+use std::{io, thread}; // use std::time::{Instant};
 
 use anyhow::Result;
 use clap::Parser;
 use cpal::traits::{DeviceTrait, HostTrait, StreamTrait};
 use cpal::{FromSample, SizedSample};
 use glicol::Engine;
-// use chrono::{DateTime, Utc};
-use crossterm::{
-    event::{self, DisableMouseCapture, EnableMouseCapture, Event, KeyCode},
-    execute,
-    terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
-};
-use ratatui::{
-    backend::{Backend, CrosstermBackend},
-    layout::{Constraint, Direction, Layout},
-    style::{Color, Style, Modifier},
-    widgets::{Block, Borders, Sparkline, Gauge},
-    text::Span,
-    symbols,
-    widgets::{Axis, Chart, Dataset, GraphType},
-    Frame, Terminal,
-};
 
-const RB_SIZE: usize = 200;
-const BLOCK_SIZE: usize = 128;
+pub const RB_SIZE: usize = 200;
+pub const BLOCK_SIZE: usize = 128;
 
 /// Glicol cli tool. This tool will watch the changes in a .glicol file.
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
 struct Args {
     /// path to the .glicol file
-    #[arg(index=1)]
+    #[arg(index = 1)]
     file: String,
 
     // Show a scope or not
     // #[arg(short, long)]
     // scope: bool,
-
     /// Set beats per minute (BPM)
     #[arg(short, long, default_value_t = 120.0)]
     bpm: f32,
@@ -83,21 +69,21 @@ fn main() -> Result<(), Box<dyn Error>> {
     let backend = CrosstermBackend::new(stdout);
     let mut terminal = Terminal::new(backend)?;
 
-    let mut ringbuf_l = [0.0; RB_SIZE];
-    let mut ringbuf_r = [0.0; RB_SIZE];
+    // let mut ringbuf_l = [0.0; RB_SIZE];
+    // let mut ringbuf_r = [0.0; RB_SIZE];
     let mut samples_l = [0.0; RB_SIZE];
     let mut samples_r = [0.0; RB_SIZE];
     let samples_index = Arc::new(AtomicUsize::new(0));
     let samples_index_clone = Arc::clone(&samples_index);
-    let index = Arc::new(AtomicUsize::new(0));
-    let index_clone = Arc::clone(&index);
-    let ptr_rb_left = Arc::new(AtomicPtr::<f32>::new( ringbuf_l.as_mut_ptr()));
-    let ptr_rb_right = Arc::new(AtomicPtr::<f32>::new( ringbuf_r.as_mut_ptr()));
-    let ptr_rb_left_clone = Arc::clone(&ptr_rb_left);
-    let ptr_rb_right_clone = Arc::clone(&ptr_rb_right);
+    // let index = Arc::new(AtomicUsize::new(0));
+    // let index_clone = Arc::clone(&index);
+    // let ptr_rb_left = Arc::new(AtomicPtr::<f32>::new(ringbuf_l.as_mut_ptr()));
+    // let ptr_rb_right = Arc::new(AtomicPtr::<f32>::new(ringbuf_r.as_mut_ptr()));
+    // let ptr_rb_left_clone = Arc::clone(&ptr_rb_left);
+    // let ptr_rb_right_clone = Arc::clone(&ptr_rb_right);
 
-    let samples_l_ptr = Arc::new(AtomicPtr::<f32>::new( samples_l.as_mut_ptr()));
-    let samples_r_ptr = Arc::new(AtomicPtr::<f32>::new( samples_r.as_mut_ptr()));
+    let samples_l_ptr = Arc::new(AtomicPtr::<f32>::new(samples_l.as_mut_ptr()));
+    let samples_r_ptr = Arc::new(AtomicPtr::<f32>::new(samples_r.as_mut_ptr()));
     let samples_l_ptr_clone = Arc::clone(&samples_l_ptr);
     let samples_r_ptr_clone = Arc::clone(&samples_r_ptr);
 
@@ -153,9 +139,14 @@ fn main() -> Result<(), Box<dyn Error>> {
     let info: String = format!("{:?} {:?}", device.name()?.clone(), config.clone());
 
     let audio_thread = thread::spawn(move || {
-        
-        let options = (ptr_rb_left_clone, ptr_rb_right_clone, index_clone, 
-            samples_l_ptr_clone, samples_r_ptr_clone, samples_index_clone, path, bpm, capacity_clone);
+        let options = (
+            samples_l_ptr_clone,
+            samples_r_ptr_clone,
+            samples_index_clone,
+            path,
+            bpm,
+            capacity_clone,
+        );
         match config.sample_format() {
             cpal::SampleFormat::I8 => run_audio::<i8>(&device, &config.into(), options),
             cpal::SampleFormat::I16 => run_audio::<i16>(&device, &config.into(), options),
@@ -178,16 +169,13 @@ fn main() -> Result<(), Box<dyn Error>> {
     let tick_rate = Duration::from_millis(16);
     let res = run_app(
         &mut terminal,
-        tick_rate, 
-        ptr_rb_left,
-        ptr_rb_right,
-        index,
+        tick_rate,
         samples_l_ptr,
         samples_r_ptr,
         samples_index,
         // scope,
         info,
-        capacity
+        capacity,
     );
 
     // restore terminal
@@ -198,89 +186,42 @@ fn main() -> Result<(), Box<dyn Error>> {
         DisableMouseCapture
     )?;
     terminal.show_cursor()?;
-
     if let Err(err) = res {
         println!("{:?}", err)
     }
-
     audio_thread.join().unwrap();
-
     Ok(())
-}
-
-fn run_app<B: Backend>(
-    terminal: &mut Terminal<B>, 
-    tick_rate: Duration, 
-    left: Arc<AtomicPtr<f32>>, 
-    right: Arc<AtomicPtr<f32>>,
-    index: Arc<AtomicUsize>,
-    samples_l_ptr: Arc<AtomicPtr<f32>>, 
-    samples_r_ptr: Arc<AtomicPtr<f32>>,
-    sampels_index: Arc<AtomicUsize>,
-    // use_scope: bool,
-    info: String,
-    capacity: Arc<AtomicU32>,
-    // right: Arc<AtomicPtr<f32>>
-) -> io::Result<()> {
-    let mut last_tick = Instant::now();
-
-    loop {
-        if true {
-            terminal.draw(|f| ui2(f, &samples_l_ptr, &samples_r_ptr, &sampels_index, &info, &capacity))?;
-        } else {
-            terminal.draw(|f| ui(f, &left, &right, &index))?;
-        }
-        
-        let timeout = tick_rate
-            .checked_sub(last_tick.elapsed())
-            .unwrap_or_else(|| Duration::from_secs(0));
-
-        if crossterm::event::poll(timeout)? {
-            if let Event::Key(key) = event::read()? {
-                if let KeyCode::Esc = key.code {
-                    return Ok(());
-                }
-            }
-        }
-
-        if last_tick.elapsed() >= tick_rate {
-            // app.on_tick();
-            last_tick = Instant::now();
-        }
-    }
 }
 
 // , play_audio: Arc<AtomicBool>
 pub fn run_audio<T>(
-    device: &cpal::Device, 
+    device: &cpal::Device,
     config: &cpal::StreamConfig,
     options: (
-        Arc<AtomicPtr<f32>>,
-        Arc<AtomicPtr<f32>>,
-        Arc<AtomicUsize>,
+        // Arc<AtomicPtr<f32>>,
+        // Arc<AtomicPtr<f32>>,
+        // Arc<AtomicUsize>,
         Arc<AtomicPtr<f32>>,
         Arc<AtomicPtr<f32>>,
         Arc<AtomicUsize>,
         String,
         f32,
-        Arc<AtomicU32>
+        Arc<AtomicU32>,
     ),
-
 ) -> Result<(), anyhow::Error>
 where
     T: SizedSample + FromSample<f32>,
 {
+    // let ptr_rb_left_clone = options.0;
+    // let ptr_rb_right_clone = options.1;
+    // let index_clone = options.2;
+    let samples_l_ptr_clone = options.0;
+    let samples_r_ptr_clone = options.1;
+    let samples_index_clone = options.2;
+    let path = options.3;
+    let bpm = options.4;
+    let capacity = options.5;
 
-    let ptr_rb_left_clone = options.0;
-    let ptr_rb_right_clone = options.1;
-    let index_clone = options.2;
-    let samples_l_ptr_clone = options.3;
-    let samples_r_ptr_clone = options.4;
-    let samples_index_clone = options.5;
-    let path = options.6;
-    let bpm = options.7;
-    let capacity = options.8;
-    
     let mut last_modified_time = metadata(&path)?.modified()?;
 
     let sr = config.sample_rate.0 as usize;
@@ -297,13 +238,13 @@ where
     let _code_ptr = Arc::clone(&code_ptr);
     let _code_len = Arc::clone(&code_len);
     let _has_update = Arc::clone(&has_update);
-    
+
     engine.set_sr(sr);
     engine.set_bpm(bpm);
 
     let channels = 2 as usize; //config.channels as usize;
 
-    let mut prev_block: [glicol_synth::Buffer::<BLOCK_SIZE>; 2] = [glicol_synth::Buffer::SILENT; 2];
+    let mut prev_block: [glicol_synth::Buffer<BLOCK_SIZE>; 2] = [glicol_synth::Buffer::SILENT; 2];
 
     let ptr = prev_block.as_mut_ptr();
     let prev_block_ptr = Arc::new(AtomicPtr::<glicol_synth::Buffer<BLOCK_SIZE>>::new(ptr));
@@ -316,11 +257,10 @@ where
     let stream = device.build_output_stream(
         config,
         move |data: &mut [T], _: &cpal::OutputCallbackInfo| {
-
             if _has_update.load(Ordering::Acquire) {
                 let ptr = _code_ptr.load(Ordering::Acquire);
                 let len = _code_len.load(Ordering::Acquire);
-                let encoded:&[u8] = unsafe { std::slice::from_raw_parts(ptr, len) };
+                let encoded: &[u8] = unsafe { std::slice::from_raw_parts(ptr, len) };
                 let code = std::str::from_utf8(encoded).unwrap().to_owned();
                 engine.update_with_code(&code);
                 _has_update.store(false, Ordering::Release);
@@ -331,33 +271,33 @@ where
 
             let samples_left_ptr = samples_l_ptr_clone.load(Ordering::SeqCst);
             let samples_right_ptr = samples_r_ptr_clone.load(Ordering::SeqCst);
-            
+
             let start_time = Instant::now();
 
-            let mut write_samples = |block: &[glicol_synth::Buffer<BLOCK_SIZE>], sample_i: usize, i: usize| {
-
-                for chan in 0..channels {
-
-                    let samples_i = samples_index_clone.load(Ordering::SeqCst);
-                    unsafe {
-                        match chan {
-                            0 => samples_left_ptr.add(samples_i).write(block[chan][i]),
-                            1 => samples_right_ptr.add(samples_i).write(block[chan][i]),
-                            _ => panic!()
+            let mut write_samples =
+                |block: &[glicol_synth::Buffer<BLOCK_SIZE>], sample_i: usize, i: usize| {
+                    for chan in 0..channels {
+                        let samples_i = samples_index_clone.load(Ordering::SeqCst);
+                        unsafe {
+                            match chan {
+                                0 => samples_left_ptr.add(samples_i).write(block[chan][i]),
+                                1 => samples_right_ptr.add(samples_i).write(block[chan][i]),
+                                _ => panic!(),
+                            };
                         };
-                    };
 
-                    samples_index_clone.store( (samples_i + 1) % 200, Ordering::SeqCst);
+                        samples_index_clone.store((samples_i + 1) % 200, Ordering::SeqCst);
 
-                    rms[chan] += block[chan][i].powf(2.0);
-                    let value: T = T::from_sample(block[chan][i]);
-                    data[sample_i*channels + chan] = value;
-                }
-            };
+                        rms[chan] += block[chan][i].powf(2.0);
+                        let value: T = T::from_sample(block[chan][i]);
+                        data[sample_i * channels + chan] = value;
+                    }
+                };
 
             let ptr = prev_block_ptr.load(Ordering::Acquire);
             let len = prev_block_len.load(Ordering::Acquire);
-            let prev_block: &mut [glicol_synth::Buffer<BLOCK_SIZE>] = unsafe { std::slice::from_raw_parts_mut(ptr, len) };
+            let prev_block: &mut [glicol_synth::Buffer<BLOCK_SIZE>] =
+                unsafe { std::slice::from_raw_parts_mut(ptr, len) };
 
             let mut writes = 0;
 
@@ -374,55 +314,56 @@ where
                         write_samples(block, writes, i);
                         writes += 1;
                     }
-            } else {
-
-                let e = block_step - writes;
-                for i in 0..e {
-                    write_samples(block, writes, i);
-                    writes += 1;
+                } else {
+                    let e = block_step - writes;
+                    for i in 0..e {
+                        write_samples(block, writes, i);
+                        writes += 1;
+                    }
+                    let mut i = 0;
+                    for buffer in prev_block.iter_mut() {
+                        buffer.copy_from_slice(&block[i]);
+                        i += 1;
+                    }
+                    prev_block_pos = e;
+                    break;
                 }
-                let mut i = 0;
-                for buffer in prev_block.iter_mut() {
-                    buffer.copy_from_slice(&block[i]);
-                    i += 1;
-                }
-                prev_block_pos = e;
-                break;
             }
-        }
 
+            let elapsed_time = start_time.elapsed().as_nanos() as f32;
+            let allowed_ns = block_step as f32 * 1_000_000_000.0 / sr as f32;
+            let perc = elapsed_time / allowed_ns;
+            capacity.store(perc.to_bits(), Ordering::Release);
 
-        let elapsed_time = start_time.elapsed().as_nanos() as f32;
-        let allowed_ns = block_step as f32 * 1_000_000_000.0 / sr as f32;
-        let perc = elapsed_time / allowed_ns;
-        capacity.store( perc.to_bits(), Ordering::Release);
+            // rms = rms
+            //     .into_iter()
+            //     .map(|x| (x / block_step as f32).sqrt())
+            //     .collect();
+            // left rms[0] right rms[1]
 
-        rms = rms.into_iter().map(|x| (x / block_step as f32).sqrt() ).collect();
-        // left rms[0] right rms[1]
+            // let ptr_l = ptr_rb_left_clone.load(Ordering::SeqCst);
+            // let ptr_r = ptr_rb_right_clone.load(Ordering::SeqCst);
 
-        let ptr_l = ptr_rb_left_clone.load(Ordering::SeqCst);
-        let ptr_r = ptr_rb_right_clone.load(Ordering::SeqCst);
-        
-        // let len = RB_SIZE;
-        let idx = index_clone.load(Ordering::SeqCst);
-        unsafe {
-            ptr_l.add(idx).write(rms[0]);
-            ptr_r.add(idx).write(rms[1]);
-        };
-        index_clone.store( (idx + 1) % RB_SIZE, Ordering::SeqCst); // from 0, 1, 2, RB_SIZE-1;
-    },
-    err_fn,
-    None,
-)?;
-stream.play()?;
+            // let len = RB_SIZE;
+            // let idx = index_clone.load(Ordering::SeqCst);
+            // unsafe {
+            //     ptr_l.add(idx).write(rms[0]);
+            //     ptr_r.add(idx).write(rms[1]);
+            // };
+            // index_clone.store((idx + 1) % RB_SIZE, Ordering::SeqCst); // from 0, 1, 2, RB_SIZE-1;
+        },
+        err_fn,
+        None,
+    )?;
+    stream.play()?;
 
-loop {
-    std::thread::sleep(Duration::from_millis(8));
-    let modified_time = metadata(&path)?.modified()?;
+    loop {
+        std::thread::sleep(Duration::from_millis(8));
+        let modified_time = metadata(&path)?.modified()?;
 
-    if modified_time != last_modified_time || has_update.load(Ordering::SeqCst) {
-        last_modified_time = modified_time;
-        let file = File::open(&path)?;
+        if modified_time != last_modified_time || has_update.load(Ordering::SeqCst) {
+            last_modified_time = modified_time;
+            let file = File::open(&path)?;
             let reader = BufReader::new(file);
             code = "".to_owned();
             for line in reader.lines() {
@@ -437,173 +378,12 @@ loop {
             // println!("\n// utc time: {} \n", datetime.format("%Y-%m-%d %H:%M:%S").to_string());
             // println!("{}", code);
             // println!("```");
-            code_ptr.store(unsafe {code.as_bytes_mut().as_mut_ptr() }, Ordering::SeqCst);
+            code_ptr.store(
+                unsafe { code.as_bytes_mut().as_mut_ptr() },
+                Ordering::SeqCst,
+            );
             code_len.store(code.len(), Ordering::SeqCst);
             has_update.store(true, Ordering::SeqCst);
         }
     }
-}
-
-fn ui(
-    f: &mut Frame,
-    left: &Arc<AtomicPtr<f32>>,
-    right: &Arc<AtomicPtr<f32>>,
-    index: &Arc<AtomicUsize>,
-    // use_scope: bool
-) {
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .margin(2)
-        .constraints([Constraint::Percentage(50), Constraint::Percentage(50)].as_ref())
-        .split(f.size());
-
-    let mut data: [f32; RB_SIZE] = [0.0; RB_SIZE];
-    let mut data2: [f32; RB_SIZE] = [0.0; RB_SIZE];
-    let ptr = left.load(Ordering::SeqCst);
-    let ptr2 = right.load(Ordering::SeqCst);
-    let mut idx = index.load(Ordering::SeqCst); // let's say 20, while RB_size is 50
-    for i in 0..RB_SIZE {
-        let value = unsafe { ptr.add(idx).read() };
-        data[RB_SIZE-1-i] = value;
-        let value = unsafe { ptr2.add(idx).read() };
-        data2[RB_SIZE-1-i] = value;
-        if idx == 0 {
-            idx = RB_SIZE - 1;// read from the tail
-        } else {
-            idx -= 1;
-        }
-    }
-
-    let leftvec = data.iter().map(|&x| (x * 100.0) as u64).collect::<Vec<u64>>();
-    let rightvec = data2.iter().map(|&x| (x * 100.0) as u64).collect::<Vec<u64>>();
-
-    let sparkline = Sparkline::default()
-        .block(
-            Block::default()
-            .title("Left").borders(Borders::ALL)
-        )
-        .data(&leftvec)
-        .style(Style::default().fg(Color::Blue));
-    f.render_widget(sparkline, chunks[0]);
-
-    let sparkline = Sparkline::default()
-        .block(
-            Block::default()
-            .title("Right").borders(Borders::ALL)
-        )
-        .data(&rightvec)
-        .style(Style::default().fg(Color::Red));
-    f.render_widget(sparkline, chunks[1]);
-}
-
-fn ui2(
-    f: &mut Frame,
-    samples_l: &Arc<AtomicPtr<f32>>, // block step length
-    samples_r: &Arc<AtomicPtr<f32>>,
-    frame_index: &Arc<AtomicUsize>,
-    info: &str,
-    capacity: &Arc<AtomicU32>
-    // use_scope: bool
-) {
-
-    let mut data = [0.0; RB_SIZE];
-    let mut data2 = [0.0; RB_SIZE];
-    let ptr = samples_l.load(Ordering::Acquire);
-    let ptr2 = samples_r.load(Ordering::Acquire);
-
-    let mut idx = frame_index.load(Ordering::Acquire);
-
-    for i in 0..RB_SIZE {
-        data[RB_SIZE-1-i] = unsafe { ptr.add(idx).read() };
-        data2[RB_SIZE-1-i] = unsafe { ptr2.add(idx).read() };
-        if idx == 0 {
-            idx = RB_SIZE - 1;// read from the tail
-        } else {
-            idx -= 1;
-        }
-    }
-
-    let left: Vec<(f64, f64)> = data.into_iter().enumerate().map(|(x, y)| (x as f64, y as f64)).collect();
-    let right: Vec<(f64, f64)> = data2.into_iter().enumerate().map(|(x, y)| (x as f64, y as f64)).collect();
-
-    let size = f.size();
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Percentage(10),
-            Constraint::Percentage(90)
-        ].as_ref())
-        .split(size);
-
-
-    let cap = capacity.load(Ordering::Acquire);
-    let portion = f32::from_bits(cap).clamp(0.0, 1.0);
-    // print!(" cap {:?}, portion {:?}", cap, portion);
-
-    let label = Span::styled(
-        format!("{:.2}%", portion * 100.0),
-        Style::default()
-            .fg(Color::White)
-            .add_modifier(Modifier::ITALIC | Modifier::BOLD),
-    );
-
-    let gauge = Gauge::default()
-    .block(Block::default().title("Render Capacity").borders(Borders::ALL))
-    .gauge_style(Style::default().fg(Color::Green))
-    .ratio(portion as f64)
-    .label(label)
-    .use_unicode(true);
-    f.render_widget(gauge, chunks[0]);
-
-    let x_labels = vec![
-        Span::styled(
-            format!("[0, 200]"),
-            Style::default().add_modifier(Modifier::BOLD),
-        ),
-    ];
-    let datasets = vec![
-        Dataset::default()
-            .name("left")
-            .marker(symbols::Marker::Braille)
-            .graph_type(GraphType::Line)
-            .style(Style::default().fg(Color::Blue))
-            .data(&left),
-        Dataset::default()
-            .name("right")
-            .marker(symbols::Marker::Braille)
-            .graph_type(GraphType::Line)
-            .style(Style::default().fg(Color::Red))
-            .data(&right),
-    ];
-
-    let chart = Chart::new(datasets)
-        .block(
-            Block::default()
-                .title(Span::styled(
-                    info.replace("SupportedStreamConfig", ""),
-                    Style::default()
-                        .fg(Color::Cyan)
-                        .add_modifier(Modifier::BOLD),
-                ))
-                .borders(Borders::NONE),
-        )
-        .x_axis(
-            Axis::default()
-                .title("X Axis")
-                .style(Style::default().fg(Color::Gray))
-                .labels(x_labels)
-                .bounds([0., 200.]),
-        )
-        .y_axis(
-            Axis::default()
-                .title("Y Axis")
-                .style(Style::default().fg(Color::Gray))
-                .labels(vec![
-                    Span::styled("-1", Style::default().add_modifier(Modifier::BOLD)),
-                    Span::raw("0"),
-                    Span::styled("1", Style::default().add_modifier(Modifier::BOLD)),
-                ])
-                .bounds([-1., 1.]),
-        );
-    f.render_widget(chart, chunks[1]);
 }
